requirements.txt
gunicorn
dash
plotly
pandas
dash_daq
uniswap-python
web3
termcolor
rich
fire
mexc-sdk
wallstreet
beautifulsoup4
